Fixed SQLGenericRepository UpdateOrCreateByIdentifier
Fixed the method on SQLGenericRepository named UpdateOrCreateByIdentifier. The problem was that the ID was not set properly on incoming data (this was done deliberately). The fix was simple. Look the item up in the database by some predicate (passed) and use this found record to update the ID of the passed item to the correct ID.